requirements.txt
streamlit>=1.38.0
google-generativeai>=0.8.3
python-dotenv>=1.0.1
PyMuPDF>=1.24.10
chromadb>=0.5.5
sentence-transformers>=5.3.0
